REL-4928: SCS memory overflow fix
Mirror commits from scs_cp REL-4928 branch:
- Comment out EPICS_TS_NTP_CONF_FILE epicsEnvSet
- tcsSad.sch: change ewaves.sbRtt FTVL from DOUBLE to FLOAT
- tcsSad.sch: change highSpeed2 FTVA and FTVB from DOUBLE to FLOAT
- tcsSad.sch: shrink highSpeed2 NOVD..NOVO outputs from 4000 to 1
- control.h: change SynchroStatus.rtt from double to float
- guide.c: convert sbRttHS array, valb scalar, and waveform pointers
  from double to float; treat valc as long (nsnrDiff)

+++ scs-mk-iocApp/src/guide.c
@@ -2130,11 +2130,11 @@ long highSpeed2 (struct genSubRecord *pgsub) {
     static long hs2count = 0;
     static long hs2subcount = HS_RECORD_LENGTH-1;
 
-    static double sbRttHS[HS_RECORD_LENGTH];
+    static float sbRttHS[HS_RECORD_LENGTH];
 
     sbRttHS[hs2count] = sbStatus->rtt;              /* next value on array*/
-    *(double *) pgsub->valb = sbStatus->rtt;        /* scalar of same rtt*/
-    *(double *) pgsub->valc = sbStatus->nsnrDiff;   /* num sent - num received*/
+    *(float *) pgsub->valb = sbStatus->rtt;        /* scalar of same rtt*/
+    *(long *) pgsub->valc = sbStatus->nsnrDiff;   /* num sent - num received*/
 
 
     if (hs2subcount == 0 ) {
@@ -2142,13 +2142,13 @@ long highSpeed2 (struct genSubRecord *pgsub) {
         int numNewSamples = HS_RECORD_LENGTH - numOldSamples;
 
         /*Input pointers */
-        double *sbRttIn = sbRttHS; /*pointer to array for sbRttHS[length] */
+        float *sbRttIn = sbRttHS; /*pointer to array for sbRttHS[length] */
 
          /*Output pointers */
-        double *sbRttOut         = (double *)pgsub->vala;  /* pointer to write the entire array (waveform) to*/
+        float *sbRttOut         = (float *)pgsub->vala;  /* pointer to write the entire array (waveform) to*/
 
-        memcpy (sbRttOut, (double *)(sbRttIn+hs2count+1),  numOldSamples*sizeof (double));
-        memcpy ((double *)(sbRttOut+numOldSamples), (double *)(sbRttIn),  numNewSamples*sizeof (double));
+        memcpy (sbRttOut, (float *)(sbRttIn+hs2count+1),  numOldSamples*sizeof (float));
+        memcpy ((float *)(sbRttOut+numOldSamples), (float *)(sbRttIn),  numNewSamples*sizeof (float));
     } else {
         hs2subcount--;
     }
